+++ noodles-bed/src/record/name.rs
@@ -1,8 +1,6 @@
 //! BED record name.
 
-use std::{error, fmt, ops::Deref, str::FromStr};
-
-const MAX_LENGTH: usize = 255;
+use std::{fmt, ops::Deref, str::FromStr};
 
 /// A BED record name.
 #[derive(Clone, Debug, Eq, PartialEq)]
@@ -22,46 +20,24 @@ impl fmt::Display for Name {
     }
 }
 
-/// An error returned when a raw BED record name fails to parse.
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub enum ParseError {
-    /// The input is empty.
-    Empty,
-    /// The input is invalid.
-    Invalid,
-}
-
-impl error::Error for ParseError {}
-
-impl fmt::Display for ParseError {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            Self::Empty => f.write_str("empty input"),
-            Self::Invalid => f.write_str("invalid input"),
-        }
-    }
-}
-
 impl FromStr for Name {
-    type Err = ParseError;
+    type Err = ();
 
     fn from_str(s: &str) -> Result<Self, Self::Err> {
-        if s.is_empty() {
-            Err(ParseError::Empty)
-        } else if !is_valid_name(s) {
-            Err(ParseError::Invalid)
-        } else {
-            Ok(Self(s.into()))
-        }
+        Ok(Self(s.into()))
     }
 }
 
-fn is_valid_name_char(c: char) -> bool {
-    matches!(c, ' '..='~')
+impl From<&str> for Name {
+    fn from(s: &str) -> Self {
+        Self::from(s.to_owned())
+    }
 }
 
-fn is_valid_name(s: &str) -> bool {
-    s.len() <= MAX_LENGTH && s.chars().all(is_valid_name_char)
+impl From<String> for Name {
+    fn from(s: String) -> Self {
+        Self(s)
+    }
 }
 
 #[cfg(test)]
@@ -76,13 +52,15 @@ mod tests {
 
     #[test]
     fn test_from_str() {
+        const MAX_LENGTH: usize = 255;
+
         assert_eq!("ndls1".parse(), Ok(Name(String::from("ndls1"))));
         assert_eq!(" ~".parse(), Ok(Name(String::from(" ~"))));
 
-        assert_eq!("".parse::<Name>(), Err(ParseError::Empty));
-        assert_eq!("🍜".parse::<Name>(), Err(ParseError::Invalid));
+        assert_eq!("".parse::<Name>(), Ok(Name(String::new())));
+        assert_eq!("🍜".parse::<Name>(), Ok(Name(String::from("🍜"))));
 
         let s = "n".repeat(MAX_LENGTH + 1);
-        assert_eq!(s.parse::<Name>(), Err(ParseError::Invalid));
+        assert_eq!(s.parse::<Name>(), Ok(Name::from(s)));
     }
 }